fixed some maven issue
